+++ Fulcrum.pro
@@ -397,6 +397,16 @@ HEADERS += \
 # Robin Hood unordered_flat_map implememntation (single header and MUCH more efficient than unordered_map!)
 HEADERS += robin_hood/robin_hood.h
 
+# Tessil hat-trie implememntation (header only, supports prefix search and fast serialization/deserialization!)
+HEADERS += \
+    tsl/htrie_set.h \
+    tsl/htrie_map.h \
+    tsl/htrie_hash.h \
+    tsl/array-hash/array_set.h \
+    tsl/array-hash/array_map.h \
+    tsl/array-hash/array_growth_policy.h \
+    tsl/array-hash/array_hash.h
+
 RESOURCES += \
     resources.qrc
